Added Chat [Say, Yell, Whisper] Fixed Name Query Players now use correct model Clean up

+++ MilkShake/MilkShake/Communication/Outgoing/World/PSNameQuery.cs
@@ -2,6 +2,7 @@
 using Milkshake.Network;
 using Milkshake.Tools.Database;
 using Milkshake.Tools.Database.Tables;
+using Milkshake.Communication.Outgoing.World.Update;
 
 namespace Milkshake.Communication.Outgoing.World
 {
@@ -9,9 +10,9 @@ namespace Milkshake.Communication.Outgoing.World
     {
         public PSNameQuery(Character character) : base(Opcodes.SMSG_NAME_QUERY_RESPONSE)
         {
-            Write((uint)character.GUID);
+            Write((ulong)character.GUID);
             Write(Encoding.UTF8.GetBytes(character.Name + '\0'));
-            Write(0);
+            Write((byte)0); // realm name for cross realm BG usage
             Write((uint)character.Race);
             Write((uint)character.Gender);
             Write((uint)character.Class);
